+++ src/renderer/transform.hpp
@@ -6,9 +6,9 @@ namespace game2d {
 
 struct TransformComponent
 {
-  glm::ivec3 position{ 0, 0, 0 };
+  glm::vec3 position{ 0, 0, 0 };
   glm::vec3 rotation_radians = { 0, 0, 0 };
-  glm::ivec3 scale{ 0, 0, 0 };
+  glm::vec3 scale{ 0, 0, 0 };
 };
 
 } // namespace game2d
